refactor: move Venue
