+++ .env.example
@@ -1,5 +1,6 @@
 BETTER_AUTH_SECRET=
 MAPS_API_KEY=
+RESEND_API_KEY=
 
 DATABASE_URL=
